Fri, 17:45, 5 files, Chris-M1-iMac
 M Alfred.alfredpreferences/workflows/user.workflow.1E686C4E-854B-4D9F-B761-EF7F221841A8/info.plist
 M Alfred.alfredpreferences/workflows/user.workflow.1E686C4E-854B-4D9F-B761-EF7F221841A8/restart-espanso.sh
 M Alfred.alfredpreferences/workflows/user.workflow.1E686C4E-854B-4D9F-B761-EF7F221841A8/restart-hammerspoon.sh
 M "Cron Jobs/frequent.log"
 M espanso/match/_shell.yml

+++ Alfred.alfredpreferences/workflows/user.workflow.1E686C4E-854B-4D9F-B761-EF7F221841A8/restart-hammerspoon.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/env zsh
 
-killall "Hammerspoon"
-sleep 1
+if pgrep -x "Hammerspoon" > /dev/null; then
+	killall "Hammerspoon"
+	sleep 1
+fi
+
 open -a "Hammerspoon"
